Move helper functions into own file, inline auth header encoding, add how_many_function property
