+++ lib/claim.js
@@ -1,43 +1,6 @@
-const querystring = require('querystring');
 const url = require('url');
 const _ = require('lodash');
-
-const content = (vars) => {
-
-  const vendor = (vars.trustedform && vars.trustedform.vendor) ?
-    vars.trustedform.vendor :
-    vars.source.name;
-
-  const params = {
-    reference: `https://next.leadconduit.com/events/${vars.lead.id}`,
-    vendor
-  };
-
-  if (vars.trustedform && vars.trustedform.scan_required_text) {
-    // Convert scan_required_text to Array if it isn't already
-    if (!(vars.trustedform.scan_required_text instanceof Array)) {
-      vars.trustedform.scan_required_text = [ vars.trustedform.scan_required_text ];
-    }
-    params['scan[]'] = vars.trustedform.scan_required_text;
-  }
-
-
-  if (vars.trustedform && vars.trustedform.scan_forbidden_text) {
-    if (!(vars.trustedform.scan_forbidden_text instanceof Array)) {
-      vars.trustedform.scan_forbidden_text = [ vars.trustedform.scan_forbidden_text ];
-    }
-    params['scan![]'] = vars.trustedform.scan_forbidden_text;
-  }
-
-  if (vars.lead.email) params.email = vars.lead.email.toString();
-  if (vars.lead.phone_1) params.phone_1 = vars.lead.phone_1.toString();
-  if (vars.lead.phone_2) params.phone_2 = vars.lead.phone_2.toString();
-  if (vars.lead.phone_3) params.phone_3 = vars.lead.phone_3.toString();
-
-  return querystring.encode(params);
-};
-
-const encodeAuthentication = (apiKey) => `Basic ${new Buffer(`API:${apiKey}`).toString('base64')}`;
+const { content, ageInSeconds, timeOnPageInSeconds, formatScanReason, countScannedRequiredStrings } = require('./helpers.js');
 
 const request = (vars) => {
 
@@ -55,7 +18,7 @@ const request = (vars) => {
     method: 'POST',
     headers: {
       Accept:        'application/json',
-      Authorization:  encodeAuthentication(apiKey),
+      Authorization:  `Basic ${new Buffer(`API:${apiKey}`).toString('base64')}`,
       'Content-Type': 'application/x-www-form-urlencoded'
     },
     body: content(vars)
@@ -74,31 +37,12 @@ request.variables = () => [
   { name: 'lead.phone_3', type: 'string', required: false, description: `Lead phone 3 that will be fingerprinted, defaults to the lead's "Phone 3" field` }
 ];
 
-
 const validate = (vars) => {
   if (!vars.lead.trustedform_cert_url) return 'TrustedForm cert URL must not be blank';
 
   // Cert URLs should begin with 'https://cert.trustedform.com/'
   const tfRegex = /^https:\/\/cert.trustedform.com\//;
   if (!tfRegex.test(vars.lead.trustedform_cert_url)) return 'TrustedForm cert URL must be valid';
-};
-
-
-const ageInSeconds = (event) => {
-  const timeOnPage = timeOnPageInSeconds(event.cert.event_duration) || 0;
-  const certAge = (new Date(event.created_at) - new Date(event.cert.created_at)) / 1000;
-  return Math.round(certAge - timeOnPage);
-};
-
-const timeOnPageInSeconds = (event_duration) => {
-  if (!event_duration || isNaN(event_duration)) return null;
-  return Math.round(parseInt(event_duration) / 1000);
-};
-
-const formatScanReason = (scannedFor, textArray)=> {
-  const matches = textArray.filter(t => scannedFor.indexOf(t) >= 0);
-  // sort text items for uniformity of reason across leads, and (arbitrarily) limit in case of long scan text
-  return `${matches.length}: '${matches.sort().join(', ').substr(0, 255)}'`;
 };
 
 const response  = (vars, req, res) => {
@@ -108,15 +52,16 @@ const response  = (vars, req, res) => {
   try {
     event = JSON.parse(res.body);
   } catch(e) {
-    // Once the new TF cert url is established, this can be refactored out and handled during validation.
-    // ie: invalidate the request if the url doesn't contain a valid hash
-    if (res.status === 404) {
-      event.message = res.body;
-    } else {
-      event.message = 'unable to parse response';
-    }
+    /*
+      The TF API sends back a string (non-JSON) response if given an
+      invalid (or non-existent) hash. Historically, requests with invalid hashes
+      have been caught by `validate()` before being sent. However, cert url validation
+      has been changed in anticipation of TF API changes. Once said changes
+      have been made, this bit can probably be refactored out and handled during validation.
+      ie: invalidate the request if the url doesn't contain a valid hash
+    */
+    event.message = (res.status === 404) ? res.body : 'unable to parse response';
   }
-
 
   if (res.status === 201 && event.cert) {
     let hosted_url = event.cert.parent_location || event.cert.location;
@@ -155,6 +100,9 @@ const response  = (vars, req, res) => {
       },
       duration: res.headers['X-Runtime']
     };
+
+    const requiredStrings = _.get(vars, 'trustedform.scan_required_text', null);
+    if (requiredStrings) appended.scans.how_many_required_found = countScannedRequiredStrings(requiredStrings, appended.scans.found); // 'all', 'some', or 'none'
 
     if (event.warnings) {
       if (event.warnings.some((warning) => warning === 'string not found in snapshot')) {
@@ -213,7 +161,6 @@ response.variables = () => [
   { name: 'scans.not_found', type: 'array', description: 'Required scan terms not found in the claim'},
   { name: 'duration', type: 'number', description: 'The number of seconds the API call took, according to TrustedForm'}
 ];
-
 
 module.exports = {
   validate,
